LU-18876 build: copy-builtin
It's useful, for the upstreaming effort, distribution packagers,
and general development, to be able to build the Lustre kernel
modules natively as part of the Linux build. Some projects
(such as openZFS) have a script to automatically port their
kernel modules to the Linux tree. Lustre should have the same.

This patch implements such a script.

This is WIP.

Test-Parameters: ignore
Signed-off-by: Timothy Day <timday@amazon.com>
Change-Id: I8fdbddd0d522f848541ae80f4873f8af1c957dc4