LU-19162 lnet: implement simple LNet block driver
Implement a simple block device on-top of LNet.

This is implemented as two drivers: lnet_host.ko (client) and
lnet_target.ko (server). The target allocates an array of
pages and registers them as passive RDMAs. The host registers
a block device and devices block requests to the target as
LNetGet() and LNetPut(). There is a one-to-one mapping between
O_DIRECT block IOs and LNet network requests.

The LNet block driver aims to implement the simplest LNet
service possible. With some additional optimization, this
could serve as an alternative to LNet selftest.

Test-Parameters: ignore
Signed-off-by: Timothy Day <timday@amazon.com>
Change-Id: Ic8968d37a76021939dcf082191598d7ca23c738a