LU-18687 lnet: move debug to LNet
Lustre debugging code will find a new home in LNet.
This code belongs to the LNet and Lustre drivers.
Since placing this in upstream lib/ is unlikely to
be accepted, we place it in LNet. Drivers (like Lustre
or LNet Selftest) that stack on-top of LNet can opt to
use these features as well.

Test-Parameters: trivial
Signed-off-by: Timothy Day <timday@amazon.com>
Signed-off-by: James Simmons <jsimmons@infradead.org>
Change-Id: I7159d602236011ac4d57c25605ecf1ba608e395d